Added Names to Proxy
Added a JSON which sits on the proxy and maps uIDs to names.

All tickets returned now have assigneeName attached to them for the front end

Added endpoint functionality for logging in and getting all users

Updating or adding a ticket with a assignee that is not in the database results in a 400 status now.

I updated the past JSON to only include valid assignees for the tickets.